requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==2.0.3
pandas==2.1.1
python-multipart==0.0.6
email-validator==2.0.0
PyJWT==2.8.0